+++ src/components/records/agRecordTable.vue
@@ -20,7 +20,7 @@ import { AgGridVue } from "ag-grid-vue";
 import {
   MappableCellRenderer,
   SelectObjectCellEditor
-} from "@/ag-grid-components/custom-renderers";
+} from "@/components/ag-grid/custom-renderers";
 
 export default {
   name: "agListTable",
